+++ src/main/java/services/DroneApi/DroneApiService.java
@@ -1,8 +1,8 @@
-package main.java.api;
+package main.java.services.DroneApi;
 
-import main.java.api.dtos.*;
 import com.google.gson.Gson;
-import main.java.api.exceptions.DroneApiException;
+import main.java.exceptions.DroneApiException;
+import main.java.services.DroneApi.dtos.*;
 
 import java.net.URI;
 import java.net.http.HttpClient;
